hw/nxagent/Rootless.c: Type cast type Atom to type long unsigned int before handing it over to %lu in fprintf().
 Amends the following compiler warnings:

 ```
 Rootless.c: In function ‘nxagentExportProperty’:
 Rootless.c:775:21: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 3 has type ‘Atom’ [-Wformat=]
                      "Property [%lu] too long.\n", propertyX);
                      ^
 ```

+++ nx-X11/programs/Xserver/hw/nxagent/Rootless.c
@@ -772,7 +772,7 @@ int nxagentExportProperty(pWin, property, type, format, mode, nUnits, value)
       {
         #ifdef WARNING
         fprintf(stderr, "nxagentExportProperty: WARNING! "
-                    "Property [%lu] too long.\n", propertyX);
+                    "Property [%lu] too long.\n", (long unsigned int)propertyX);
         #endif
 
         goto nxagentExportPropertyError;
